Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/streams/piping/transform-streams.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL transform-streams.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
